Refactor journey details stop components and time helpers.
