+++ src/components/sidepanel/journeyDetails/elements.js
@@ -1,6 +1,8 @@
 import styled from "styled-components";
 import TimingStopIcon from "../../../icon-time1.svg";
 import React from "react";
+import {TagButton} from "../../TagButton";
+import {Heading} from "../../Typography";
 
 export const SmallText = styled.span`
   display: block;
@@ -57,3 +59,30 @@ export const StopElementsWrapper = styled.div`
     margin-top: auto;
   }
 `;
+export const StopWrapper = styled.div`
+  padding: 0 1rem 0 0;
+  margin-left: 0.75rem;
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+`;
+export const StopContent = styled.div`
+  padding: 0 1.25rem 2rem 0.75rem;
+  width: 100%;
+`;
+export const StopHeading = styled(Heading).attrs({level: 5})`
+  margin-top: 0.2rem;
+  color: var(--dark-grey);
+  font-size: 0.875rem;
+  font-weight: normal;
+`;
+export const TimeHeading = styled.div`
+  font-size: 0.75rem;
+  color: var(--light-grey);
+  margin-bottom: 0.2rem;
+  margin-top: 1rem;
+`;
+export const StopArrivalTime = styled(TagButton)`
+  margin: 0 0 0.5rem;
+`;
+export const StopDepartureTime = styled(TagButton)``;
