+++ src/helpers/stopArrivalTimes.js
@@ -3,28 +3,34 @@ import moment from "moment-timezone";
 import {TIMEZONE} from "../constants";
 import {getAdjustedDepartureDate} from "./getAdjustedDepartureDate";
 
-export const stopArrivalTimes = (positions = [], departure, date) => {
-  const event = positions[0];
+/**
+ *
+ * @param stopPositions positions with next_stop_id = [current stop]
+ * @param stopDeparture the planned departure from [current stop]
+ * @param date selected date in YYYY-MM-DD format
+ * @returns {*}
+ */
+export const stopArrivalTimes = (stopPositions = [], stopDeparture, date) => {
+  let arrivalEvent = stopPositions[0];
 
-  if (!event || !departure || !date) {
+  if (!arrivalEvent || !stopDeparture || !date) {
     return false;
   }
 
   // Find out when the vehicle arrived at the stop
   // by looking at when the doors were opened.
   let doorDidOpen = false;
-  let arrivalEvent = event;
 
-  if (event) {
-    for (let i = 0; i < positions.length; i++) {
-      const position = positions[i];
+  if (arrivalEvent) {
+    for (let i = 0; i < stopPositions.length; i++) {
+      const position = stopPositions[i];
 
       if (doorDidOpen && !position.drst) {
-        arrivalEvent = positions[i - 1];
+        arrivalEvent = stopPositions[i - 1];
 
         // If that didn't exist, just pick the current item as a fallback.
         if (!arrivalEvent) {
-          arrivalEvent = positions[i];
+          arrivalEvent = stopPositions[i];
         }
         break;
       }
@@ -45,6 +51,7 @@ export const stopArrivalTimes = (positions = [], departure, date) => {
   return {
     arrivalEvent,
     arrivalMoment,
-    plannedArrivalMoment: getAdjustedDepartureDate(departure, date, true),
+    departure: stopDeparture,
+    plannedArrivalMoment: getAdjustedDepartureDate(stopDeparture, date, true),
   };
 };
